Add ability to alter message profiles
